+++ changelog.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Display current git branch of project in text view
 - Add menu to list git repository branches
 - Add search field for git branches menu
+- Add checkout action on branch menu
 - Draw component on definition click
 
 [unreleased]: https://github.com/ditrit/leto-modelizer/blob/main/changelog.md#unreleased